삭제 버튼 클릭시 삭제하시겠습니까 라는 모달창 구현 및 확인 버튼 클릭 후 삭제 + 홈 화면 이동 설정

+++ src/components/views/DetailView.vue
@@ -6,7 +6,9 @@
 
     <div>
       <b-link @click="deleteWriting($route.query.id)">삭제</b-link>
-      <b-link class="updateWriting" @click="updateWriting($route.query.id)">수정</b-link>
+      <b-link class="updateWriting" @click="updateWriting($route.query.id)"
+        >수정</b-link
+      >
     </div>
 
     <div class="detailContext">
@@ -41,14 +43,45 @@ export default {
 
   methods: {
     deleteWriting(id) {
-      axios.delete("/api/writings/delete/" + id).then(() => {
-        console.log("삭제완료");
-      });
+      this.$bvModal
+        .msgBoxConfirm("삭제하겠습니까?", {
+          size: "sm",
+          buttonSize: "sm",
+          okVariant: "danger",
+          okTitle: "예",
+          cancelTitle: "아니오",
+          footerClass: "p-2",
+          hideHeaderClose: false,
+          centered: true,
+          cancelCallback: () => {},
+        })
+        .then((value) => {
+          if (value) {
+            axios.delete("/api/writings/delete/" + id).then(() => {
+              this.$bvModal
+                .msgBoxOk("삭제되었습니다.", {
+                  size: "sm",
+                  buttonSize: "sm",
+                  centered: true,
+                  okTitle: "예",
+                })
+                .then(() => {
+                  this.$router.push("/");
+                });
+            });
+          } else {
+            return null;
+          }
+        });
     },
     updateWriting() {
       this.$router.push({
         path: "/update",
-        query: { title: this.titleValue, context: this.contextValue, id: this.id },
+        query: {
+          title: this.titleValue,
+          context: this.contextValue,
+          id: this.id,
+        },
       });
     },
   },
